src/requirement.txt
streamlit
requests
opencv-python-headless
pillow
numpy
matplotlib
scikit-image
scipy
exifread
tqdm
torch
torchvision
torchaudio
plotly

